requirements.txt
psycopg2-binary
streamlit
pandas
plotly==5.18.0
openpyxl==3.1.2